Refactor cluster wrappers into scripts/wrappers/ folder
