+++ microk8s-resources/wrappers/microk8s-join.wrapper
@@ -22,4 +22,4 @@ fi
 
 exit_if_no_permissions
 
-sudo -E  LD_LIBRARY_PATH=$IN_SNAP_LD_LIBRARY_PATH ${SNAP}/usr/bin/python3 ${SNAP}/scripts/cluster/join.py $@
+sudo -E  LD_LIBRARY_PATH=$IN_SNAP_LD_LIBRARY_PATH ${SNAP}/usr/bin/python3 ${SNAP}/scripts/wrappers/cluster/join.py $@
