+++ tox.ini
@@ -38,7 +38,7 @@ commands =
         --ignore=tests/unit/test_addons.py {posargs}
 
 [testenv:cluster]
-setenv = PYTHONPATH={toxinidir}/scripts/cluster
+setenv = PYTHONPATH={toxinidir}/scripts/wrappers
 commands =
     pytest -s tests/unit/cluster {posargs}
